Fix possible buffer overrun in options find_match
If a match_type is not MATCH_LONG then the match_name
pointer points to p->short_opt which is a single char.

This is passed to strlen a few lines later. This will be
a buffer overrun if short_opt is not the null char.

The fix I'm doing here is to use strnlen and pass the
match_length, which is set to 1 for the case of
a short option.