Alternate fix to possible overrun in option parsing
This is an alternate fix to the issue that was attempted
to be resolved in commit d1901df1. That fix had problems
with options that are substrings of other options. For
example, --antlr vs --antlrraw.

This fix explicitly checks for a SHORT option and does
not do the strlen test in that case. match_length will
always be 1 and the short option is length one so the
test will succeed.