Fix issue 577 and add tests for String.count.
String.count would fail to count the last instance of a substring in a string if the substring appeared twice at the end of the string. For example, "atat".count("at") would return 1 instead of 2.

The problem was caused because the loop that drives the search was terminated when the loop index was greater than or equal to the difference in length of the two strings being compared. Because of the way String.count used the String.find method, this was only a problem in situations where finding the second-to-last-string resulted in the index now being equal to the difference in length of the two strings being compared, causing the loop to terminate before finding the last occurrence of the substring.

The solution was to replace `k < j` with `k <= j` in the match search loop.

I added grey-box tests for String.count to prevent a regression on this issue and to excercise the code paths through the method.