Remove public/private output from entity documentation
All we are doing is checking to see if the first character of the name is '_'.
This is standard Pony idiom and enforced by the compiler. We can expect
that anyone seeing '_name' knows that it is private and that 'name' is public.

The removed output added little value while cluttering up the documentation.