Rewrite the code generator to be iterative
The original code generator was recursive, and walked the expression
AST to determine what types and functions should be generated. Later,
reachability determination was added to allow the correct reifications
of polymorphic functions to be generated.

This new code generator leverages the reachability work to replace
the recursive aspect entirely. Reachability is now used to determine
all of the types and functions to generate. In addition, LLVM types,
values, and debug information are all cached in the reachability
data structures to improve compile times.

As part of this work, the debug info generator has been rewritten to
use the metadata approach of LLVM 3.7+.