Use primitives for capabilities-secure networking
This is based on @dkck's work on capability-secure networking.
The idea is to use primitives as authorisation tokens, so that
higher level interfaces can be built that hold those primitives,
but can do additional validation. For example, an application
might want to limit the ports that can be listened on. To do so,
pass a class that can construct a TCPListener (by holding a
TCPListenerAuth token) but will only do so for a specified port
range.