Properly scope local `let` in trait default implementations
When a trait is applied to a type, we copy over the method body
to the new type and as part of this scope everything. This works
fine for `var` which can be redefined at will. Let is a different
story. Given that we already scoped and defined our let, when we go
to do this on the actual type, we end up with an error that we are
redefining it. What we want to be doing is to set local to undefined
and allow them to be set to defined when the expression type checking
happens in the new type.

Closes #684