Paramaterize Array.find/rfind on a comparator
Add a parameter `predicate` to the `find` and `rfind` methods of Array
that accepts a lambda returning Bool, which replaces the `is` test in
the methods. By default, to maintain current behavior, this predicate
is passed a lambda that uses an `is` test, but it can be replaced with
a test using equality or any other binary predicate.

Add tests for find and rfind.

Fixes #659