Add `ArrayValues.rewind()` method
The `ArrayValues.rewind()` method "rewinds" so that subsequents calls
to `next()` return items starting from the first item in the array.

This came as a result of a conversation with Sylvan. There was a
situation where it made sense to have `rewind()` as part of
`ArrayValues`, but we didn't think it made sense to make it part of
the `Iterator` interface.