Drop dynamic LTO detection in the build system
Previously, LTO was dynamically enabled if the required components were
found. However, there are some differences between compilers for LTO
support which led to many corner cases and fixing those would have
added a lot of complexity in the Makefile.

LTO is now disabled by default and the manual setup is explained in the
Readme. Since LTO setup is automatic on OSX, it is still enabled by
default on this platform.