Check local compatibility for intersection types
An intersection type represents multiple views on to the same
object. As such, every view in an intersection type must be
locally compatible with every other view. This check is introduced
in the flatten pass (and skipped for provides lists and constraints
as they do not represent actual object view). As a result, the
subtyping check for intersection types is simplified to the natural
form, i.e.:

  T1 <: T3 or T2 <: T3
  ---
  (T1 & T2) <: T3