Only sort program level documentation elements
The order that we put definitions in source files is more
meaningful than alphabetizing. With this change, we preserve
the alphabetizing ONLY for package level. This makes sense,
the programmer hasn't provided explicit ordering for the
packages that make up the program, but they have for the types
that make up the various packages and modules.