Print default values in documentation
Before this commit, we printed "optional" next to parameters
that came with default values. Nice, but not as useful as
knowing what that value is.

This change should had "straight forward" defaults. If the default
is a complex expression, it will probably render like shit and per
this discussion: https://pony.groups.io/g/dev/thread/724212#15,
we'll need to revisit and improve that printing. However, this commit
should work for just about everything that is currently in the
stdlib.

Closes #682