Improve method documentation format
The previous format was hard to parse.
This introduces a much easier to read & parse layout.

It features a code block of the complete method signature
that is much easier to read than what their was before.

However, code blocks can't have links in them so we have
additional sections for the parameters and return types
that can continue to have links to documentation for other
classes.