Make Travis only build on pushes to master (and PRs).
Currently Travis will build twice for PRs by committers on branches
that are in the repo - once for the branch and once for the merge.
Also, it will build anytime a branch is pushed, even if it's not
ready to PR yet. This commit stops both of these undesired behaviours.