Appveyor (#735)
* Don't assert on union/isect flatten

It's possible to hit the flatten pass on types that have already
been flattened, for example when generating an anonymous type for
a lambda. Instead of asserting that the child count is 2, the
flatten step can be skipped if the child count is greater thant 2.

* Primitive _init now takes no arguments

As pointed out by @dckc, primitive _init taking an Env as an
argument gives ambient authority to any package declaring a
primitive with an _init function. Since these functions are
meant primarily for initialising FFI libraries, access to Env
isn't needed. If a use case appears where some part of Env is
needed, this can be revisited.

* Appveyor YAML

Use a .appveyor.yml file instead of configuring the build on the
appveyor website.

* try to make ps init script work

* try to make ps init script work

* try to make ps init script work, v3

* try to make ps init script work, v4

* try to make ps init script work, v5

* try to make ps init script work, v6