Add `from_tuples()` and `from_array` to HashMap
This commit adds the constructors `from_tuples()` and `from_array` to
the HashMap class. There are probably several arguments for having
convenience functions like this, but the primary motivation for this
was the question of whether or not Pony had sugar for creating
maps. In the interest of keeping sugar to a minimum I think this is
the next best thing.

`from_tuples()` creates a HashMap from an array of tuples of key/value
pairs.

`from_array()` creates a HashMap from an array of alternating keys and
values.