Add process package that enables control of external processes [pr# 3] (#772)
* Add process package that enables control of external processes

This package gives you control over an external process. You
can spawn any process and have control over STDIN, STDOUT and
STDERR of the spawned process. All communications is done via
async IO and will not block. Any errors during operation related to
system calls are surfaced to the caller as ProcessError via the
notifier.

* Add FileExec capability to caprights and filecaps

ProcessMonitor needs to check the capabilities of the caller
before calling execve. We add a new cap named FileExec which
represents this capability.

Add: FileExec

* Add FileExec check in ProcessMonitor constructor

We are now checking if the caller granted us the capability
to execve the requested command.