Machine words are subtypes of some traits and interfaces, but have a different binary representation from an object. Previously, the compiler attempted to detect when machine word representation would make method subtyping result in unusable representations (e.g. passing a 4 byte integer where a memory pointer was expected, etc.). This failed in the presence of complex generics.This change instead allows machine words subtyping to work like any other types, and generates forwarding functions with mangled names that correctly autobox machine word types when (and only when) it is required.Fixes #561 and #563
Machine words are subtypes of some traits and interfaces, but have a different binary representation from an object. Previously, the compiler attempted to detect when machine word representation would make method subtyping result in unusable representations (e.g. passing a 4 byte integer where a memory pointer was expected, etc.). This failed in the presence of complex generics.

This change instead allows machine words subtyping to work like any other types, and generates forwarding functions with mangled names that correctly autobox machine word types when (and only when) it is required.

Fixes #561 and #563.