Rename identityof to digestof
The name of the identityof operator was misleading. It did not, and
could not, return a value that was unique for every possible
expression, because it returns a U64, which cannot encode the
identity of either 128 bit integers or tuples.

This is fine, since the purpose of the operator is not to return an
identity. The actual purpose is to return a number derived from the
identity that is suitable for hashing: in other words, a digest.