Improve delegate example code
The old delegate example code didn't really demonstrate what a
delegate did, which was confusing for me when I initially started to
try to use delegates. It creates several objects, some of which are
used as delegates, but then calls those objects directly. To
demonstrate delegation, the delegate should be set to one of those
classes, and then the delegating object should be called to
demonstrate that it is delegating the call to the delegate. This
commit implements that behavior.