Generating tuple return value with tuple bodies
Previously, this line of code always switch to the return type for
gen_assign_cast if the return type was a tuple. This is needed in
order to codegen the right LLVM return type. However, if the body
type is also a tuple, then the body type, rather than the result
type, should be passed to gen_assign_cast so that the correct
boxed type can be chosen by gen_box if necessary.