Use LLVM i8 for Bool instead of i1
The LLVM load-combine pass can produce invalid results when
combining i1 loads. Previously, it was believed this only
happened on OSX, so the load-combine pass was disabled on OSX.
However, a Linux test case has been found.

The performance cost of disabling load-combine is high. Switching
to representing Bool as i8 instead of i1 has no runtime cost, and
may actually improve performance depending on how LLVM was storing
sequences of i8 in structures.

The load-combine pass is now re-enabled on all platforms.