Memory leak when iso send is resent as val
Previously, something resent as immutable was fully traced. This
was to protect against a premature free on release when the object
was not reachable by the owner. However, it resulted in an off-by-
one error in the RC and a memory leak.

Now, objects resent as immutable need not be fully traced, which is
a performance optimisation as well as removing the memory leak. As
a result, reachability on locally owned objects that have been sent
in messages is no longer tracked, so a release message can't result
in an immediate free (a gc cycle has to happen).