String.from_array should copy _size bytes
Previously, it was copying _alloc bytes from the array, when the
array did not have _alloc bytes available. This would often succeed
due to an extra byte being read not causing a problem. However, for
a zero-length array or an array aligned right on a page boundary,
this would segfault.

Copying _size bytes always works, even for a zero-length array.