Normalize install instructions
Use '$' at start of all shell commands.
Sometimes we had nothing, other times we had '>' and
sometimes we had '$'.

I had no dog in this race. '$' seems most often used.
I prefer nothing as it is easier to copy and paste but
I am going with the majority usage for now.