Use the nice name for types when generating documentation (#872)
In generated documentation, replace the use of hygienic, generated
names for anonymous function types with a nice name based on the type
of the function. Generated names have the form of `$0$1` and are less than useful.

This change associates a nice name with function types based on the
`{(ISize):ISize}` syntax and uses that name in generated documentation.
Additionally, this change adds the reference capability for the function
to the nice name, as in `{ref(ISize):ISize} ref`.