Fix compiler crash while generating C lib (#891)
The compiler crashes while generating the header for the C library.
Particularly, the compiler crashes when the exported actor
(the actor with @ in front of the actor's name)
contains the method that returns variadic type containing None
(eg., 'T | None').

The longer explanation follows.

The following program:

  primitive T

  actor Main
    new create(env: Env) =>
      f(T)
    fun f(a: T): (T | None) =>
      a

compiles correctly when running 'ponyc'.

When we want to compile the same program as the C library ('ponyc -l')
we have to add @ symbol in front of the 'actor Main'.
So the new program becomes:

  primitive T

  actor @Main
    new create(env: Env) =>
      f(T)
    fun f(a: T): (T | None) =>
      a

Unfortunatelly, this crashes the compiler:

  # ponyc -l
  Building builtin -> /home/srdjan/prog/github/ponyc/packages/builtin
  Building . -> /home/srdjan/prog/pony/test
  Generating
   Library reachability
   Selector painting
   Data prototypes
   Data types
   Function prototypes
   Functions
   Descriptors
  ponyc: src/libponyc/ast/ast.c:678: ast_childidx: Assertion `ast != ((void *)0)' failed.
  [2]    23160 abort (core dumped)  ponyc -l

The following program is the minimal program/test
that crashes the compiler in the same way:

  primitive T

  actor @A
    fun f(a: T): (T | None) =>
        a