Set name resolution pass errors to be fatal.
This stops some more exotic forms of type alias recursion (such as
when a type aliases a union or tuple, where one of the elements has
the type alias as a type parameter), by not allowing the compiler
to continue after it has detected a recursion.

A possible downside of this change is that *all* name resolution
pass errors will be treated as fatal, meaning that users may no
longer see parallel name resolution pass errors in the output.
In practice, this is probably not a big deal, and its easier than
the alternative of trying to complicate the codebase by differentiating
between type alias recursion and other name resolution pass errors.

Resolves #901.