Set correct source file for debug info (#939)
In codegen_pushscope, use the existing frame->di_file rather than
trying to create a new one, as that can lead to incorrect results.

In make_debug_info, if the type being generated has a definition
(i.e. it is a nominal type), use the file the definition is in,
rather than whatever random source file the use-site AST node
happens to be in.