Heuristics for GC when receiving immutable objects (#942)
Remote immutable objects used to be traced in their entirety. Their
heap size was added to the local actor heap size. This was optimised
so that they need not be traced, but this resulted in the accumulated
heap size being far too low.

This change introduces a heuristic for remote immutable memory use. It
also stops accounting for remote objects in the effective local heap
size, so that only only local objects count as heap size after a GC
pass. As a result, remote objects and actors provoke GC, but never
delay it.

As part of this change, message batching on an actor is terminated
early if the actor's queue when it was scheduled is exhausted. This
prevents a batch from handling messages that the batch itself has
generated, or that arrive while a batch is in progress.