Added minimal tests for examples.
Currently the examples are not included in the CI process, this leads to
this kind of commit messages :

4eb42e2 Fix ffi example (#949) Was broken a while back.

I added a separate make target `test-examples`though I believe this
tests should be run with the full tests suite.

The package is minimal at the moment, it just tries to use all examples.
Unit tests should be added later, demonstrating some testing techniques.

Some example packages are commented at the moment as they don't build on
my laptop. I still have to figure out what is broken: the example or my
pony build chain.