Check syntax in lambdas
The `pass_syntax` compiler pass was skipping lambda functions because
the child nodes of the lambda function are all intially flagged as
AST_FLAG_PRESERVE when `pass_syntax` is run in `module_passes`; this
flag is removed in a later step. This fix calls `pass_syntax` again
after the flag is removed in order to catch any issues that were
missed on the first call to `pass_syntax`.

Fixes issue #981.