Check syntax in lambdas (#985)
The pass_syntax compiler pass was skipping lambda functions because
the child nodes of the lambda function are all intially flagged as
AST_FLAG_PRESERVE when pass_syntax is run in module_passes; this
flag is removed in a later step. This fix calls pass_syntax again
after the flag is removed in order to catch any issues that were
missed on the first call to pass_syntax.

In order for this to work correctly, `check_id` was changed to allow
any identifier that starts with a "$" character. The reasoning is
that variables with these names are only used in hygienic variables
that are created by the compiler, and in certain tests. We don't want
to disallow them, so we assume that if you use them then you know what
you are doing. This has been discussed with @sylvanc.

Once this change was made the compiler spotted several errors in test
code which had been previously overlooked. These tests were also
changed as part of this fix.