Don't assert on unknown mutable remote objects
A false assumption was made that a mutable remote object must be
known. The counter-example is that an embedded field of an iso
won't have its own foreign reference count, but can be assigned
to a field and traced, at which point an independent foreign
reference count must be generated.