Allow _final as a field name (#990)
A _final field is now allowed. A _final method must still be a fun
rather than a new or a be, and must have the correct signature for
a finaliser, to avoid a situation where a programmer believes they
have written a finaliser, but the compiler doesn't recognise it and
so the runtime doesn't call it.