Ignore SIGPIPE (#999)
Set SIG_IGN on SIGPIPE. Previously, individual sockets were marked
with SO_NOSIGPIPE on OSX and FreeBSD, and individual calls to send
were flagged with MSG_NOSIGNAL on Linux. This only works with calls
to send, and not with write or writev.

Now, SIGPIPE is ignored by default. If an application wants access
to SIGPIPE, it is still available via the signals package.