Reserve chunks (#1013)
* Allow for reserving of chunks in WriteBuffer

When using the WriteBuffer, I find myself knowing
how many chunks I intend to use. With this change,
anyone in a similar situation can preallocate the
correct number of chunks for the buffer.

Note as this is currently written, this needs to be
reset after each call to done() as done() resets
chunks back to Array[ByteSeq].

* Allow for reserving of chunks in WriteBuffer

When using the WriteBuffer, I find myself knowing
how many chunks I intend to use. With this change,
anyone in a similar situation can preallocate the
correct number of chunks for the buffer.

Note as this is currently written, this needs to be
reset after each call to done() as done() resets
chunks back to Array[ByteSeq].