Add write buffering and asio events for writes to STDIN of the child
We now register an asio event on the STDIN file descriptor.
We write as much as possible. If we can't write everything we
store the ByteSeq in a list together with an offset and set the
_stdin_writeable flag to false.
Once the OS signals STDIN is writeable again we try to write the
chunks stored in the list starting with the head chunk.