Make JSON pretty printing optional
Pretty printing JSON creates a slightly larger data structure than
a non-pretty printed version. When dealing with large aggregates,
this can result in large differences that can have a noticeable
impact on performance when sending over the network.

This change makes pretty printing optional in a way that maintains
compatibility with the existing API.