Increase timeout for TestStdinWriteBuf test to 20 seconds
The build server might run these tests on a resource constraint
node. In this scenario we push data into STDIN of the forked process
via a connected pipe and consume it from it's STDOUT connected pipe.
The time to complete this test might vary greatly depending on the
pipe capacity and cpu clicks received to _copy_ the data through two
pipes.