Fix incorrect use of the `len` variable in regex package replace function (#1033)
In `regex.pony` line 132, we rely on the `len` variable when
`PCRE2_ERROR_NOMEMORY` has been encountered by PCRE2. This is not
correct because we pass the pointer to `len` to `pcre2_substitute_8`
and it modifies the value (see: http://vcs.pcre.org/pcre2/code/tags/pcre2-10.21/src/pcre2_substitute.c?revision=476&view=markup
line 242) so that it is not usable for our purposes any longer.