Add `offset` to String.from_array constructor
This change adds an additional parameter to String's from_array
constructor. There are times when you want to construct a String
from an array but not the entire array. For those situations,
this will be a performance improvement.

This is a non-breaking change as existing functionality is
maintained via the default value of 0 that is set for offset
when not applied.