Improve tcp notifier performance
Sending via TCP connections with writev is incredibly slow in
the default scenario. Pony's `error` keyword while much faster
than exceptions in Java is still relatively slow. By using,
error in the default case for `sentv` then any user code that
uses `TCPConnection.writev` which calls `sentv` will be very
slow. You'd have to know that you want to override `writev`
in order to get better performance.

The simple default here is to return `data` unchanged from
`sentv` rather than throwing the error.

This does have a negative impact on the user API in that
you can no longer easily use your custom handling in `sent`
from `sentv` without having to do anything.

After discussion during a Pony Sync meeting, it was decided
that this change was the correct thing to undertake.

Sylvan has expressed interest in adding on to this commit
with some changes to how SSL with operate with this change.
As such, this change works and compiles but perhaps shouldn't
be considered complete as is.