Correctly add prefixes for hex and octal strings (#1094)
If a number was converted to a string with either a hex or octal
representation, and a precision was specified, then the resulting
string had `0`s before the hex or octal prefix. For example, `10`,
when converted to hex with a precision of 5, resulted in "00000xA";
the expected result is "0x0000A".

This is now fixed. The fix involved rearranging the way the string was
generated so that the prefix is added after extending the digits (the
string is constructed backwards and then reversed). A test has also
been added that uses different formatting options to format a string
and then checks the results.

Closes #1093