Make composite types print to mutable string
This commit adds a _show method to the JsonArray and JsonObject types.
This method is responsible for printing the string representation to a
string buffer that it receives as an argument. The method also receives
the desired indentation style, and the current level of indentation.

The indentation is managed by a new _JsonPrint._indent method, which
prints the leading newline and the indentation as many times as the
given level requires.

The _JsonPrint._string method has been updated to also receive the
buffer and will either pass the string representation of the primitive
to it, or will forward it to the _show method of JsonArray or
JsonObject.

This greatly reduces the amount of allocation that takes place.

The 'I64' and 'String' types in the _JsonPrint match expression have
also been altered to avoid unnecessary allocation. The 'F64' still
requires an allocation at the moment.

When only a single byte is added to the buffer, .push() is used instead
of .append().

Lastly, the JsonDoc.string() method now allows the user to pass the
desired indentation style.