Add LLVM type-based alias analysis metadatas
These metadatas make LLVM aware of the aliasing guarantees of the Pony
type system. It will allow a more precise alias analysis, which will
lead to better optimisations.

Metadatas are added on loads and stores of fields. Since fields of
interfaces, union types, etc. are inaccessible, only base types are
considered. Each type has its own subtree (objects of different types
can never alias) which is arranged in the following way:

                        ----R----
                       /    |    \
                     iso   box   val
                          /   \
                        ref   trn

In LLVM, an object of a given type can alias with objects of any child
or parent type. The tree is a simple transposition of the reference
capabilities rules. The only difference is val ; it isn't a child of
box here because LLVM aliasing only cares about writes. tag is left out
since it can neither be read from nor written to.