Add 'from_iso_array' constructor to String (#1132)
When you receive data in the format of Array[U8] from something like a TCPConnection,
you often want to work with it as a String and want to be able to mutate it without
making a copy. This constructor allows you to turn an iso Array into an iso String.

For example, you could receive data via TCP, turn into an iso String and then
call methods like lower_in_place on it, all without additional allocations.