Force constructor inlining when stack-allocating objects (#1144)
This is an improvement for the heap2stack optimisation pass, which
replaces heap allocations by stack allocations when possible. This
modification allows the optimiser to know that the parent object of a
given field is stack-allocated, which gives it more information to
stack-allocate the field itself.

In order to get correct nocapture info in LLVM, which greatly helps the
analysis in heap2stack, class constructors in generated code now return
void instead of their receiver.