Use existing while-else condition to exit early
This removes the early check for an empty string and replaces it with an
early exit in the existing `else` of the inner `while` loops of methods
that perform a search.

Because the `else` of those loops will only be reached when the search
pattern is empty, and will consistently do so resulting in a failed
search, it is safe and sensible to exist after the first determination
of an empty search pattern.