Revert "Add nounwind to non-partial functions (#1147)"
This reverts commit e56075d46d7d9e1d8c5e8be7ed0506ad2de98734.

Revert to fix bug #1212

Here's a program to demonstrate on OSX with LLVM 3.8.
So far it appears to not impact Linux.

```
use "collections"

actor Main
  new create(env: Env) =>
    for i in Range[U64](1, 5) do
        let r = Rectangle(i, i + 2)
        env.out.print("Width and height: " + r.dimensions())
        env.out.print("Width and height: " + r.dimensions())
        env.out.print("")
    end

class Rectangle
  let _width: U64
  let _height: U64

  new create(w: U64, h: U64) =>
    _width = w
    _height = h

  fun dimensions(): String => _width.string() + " " + _height.string()
```

When run after the revert commit, instead of correct results, you get:

```
Width and height: 1 3
Width and height: 4445790517 4445957376

Width and height: 2 4
Width and height: 4445788213 4445957376

Width and height: 3 5
Width and height: 4445788725 4445957376

Width and height: 4 6
Width and height: 4445786165 4445957376
```