Make version come from VERSION file, not git tags
The version assigned when building was coming from `git describe --tags
--always` when `.git` was present, otherwise from `VERSION`. This can
get out of sync with `VERSION` if the git tags are not up to date.

This commit makes the `tag` variable come version come exclusively from
the `VERSION` file, which then is used to set the build version.

This rendered the `setversion` command, which used the `tag` variable
to set the `VERSION`, file recursive and nonsensical, so the
`setversion` command has been removed. The `git` variable was also
removed, since it was only use was for `setversion`.

The `setversion` was also removed from the `release` command.

This resolves issue #1220.