Make `release` the default build type
Prior to this commit, the default when
running `make` was to create a debug
build. This is fine for pony developers
but pony first time users might be unaware.

Our stated guiding principle for defaults
is to use the best performing option that
wouldn't violate the principal of least
surprise. And so, this commit.