Enable LTO by default for OSX builds
Commit #1207 turned off LTO as a default. Subsequent
testing by Sendence found a ~ 10% performance
degradation for some operations with the change.

The change was made in order to get the Travis CI
builds working. Not a wonderful trade-off. This change
turns LTO back on by default on OSX and fixes our
Travis issue.

Eventually, Travis should have a compiler that doesn't
have issue with LTO and we can remove the Travis portion
of this.