Fix Travis OSX builds (#1230)
A change in Homebrew is breaking OSX builds.
This fix handles the problems we are now having
installing GMP. Specifically, we are hitting errors
were it can't be linked into /usr/local when
being installed as a dependency for various
llvm version. This ends up failing the llvm installs
and from that, the build.

This fixes by "manually" installing gmp and then
forcing the link overwrite.