Stop generating `llvm.invariant.load` intrinsic for "let" references.
Resolves #1122.

As we saw in issue #1122, the load invariant doesn't necessarily hold true
for let references in Pony. Particularly it doesn't seem to hold for `for`
loops, which are `let` in Pony semantics, but the value changes between
iterations of the loop, so the optimizer can't see them as invariant.

With more research it might be possible to exclude the `for` loop case
(and any other problematic cases) but for now the best course of action
is to remove it for all `let` references.