buffered - Added a string() method on Reader
Consume and return a null-terminated string.

Some protocols need this (e.g.
https://www.postgresql.org/docs/current/static/protocol-message-formats.html),
I had to implement this by iterating over U8s which is far from
optimal and not very user-friendly.

About the name of the function, I guess it's useless to make Reader
a Stringable, so I took the name. I can change this if needed.