Fix copy bounds for String.clone (avoid segfault for trimmed strings).
We are calling `_set(len, 0)` to set the null terminator on the new
string, so there is no reason to copy `len + 1` bytes from the original
string pointer. In fact, for trimmed strings this may not be safe -
in some situations, `len + 1` bytes may end in a chunk of memory that
we don't actually own, causing a segmentation fault.