Add line breaks for bar separated lists
Bar separated lists were causing an ugly horizontal scroll
bar in the docs. This commit fixes #1263 by including a new
line every three list elements if the separator is a bar. The
behavior is added as a boolean flag to the doc_type_list function
to be easily applied to other lists if the need arises.