Add DTrace/SystemTap support to Pony (#1324)
In this contribution we have replaced the telemetry functionality of the compiler by scripts for both technologies. It shows the clean approach of DTrace/SystemTap. Support for Windows might be limited by this change. However since there was no support for the telemetry in the Windows build tools, we thought this would be the best approach.

Currently implemented probes have a focus on replacing the events used by the telemetry functionality. We plan to add additional probes in the coming weeks, but we were hoping on possible feedback to change our approach if necessary.

Contributors to this addition are: @mkdubik, @DarkLog1x, and me