Verify TCP Connections are open before allowing data to be written.
Prevent subtle bugs and potential runaway memory usage that can be introduced by being unfamiliar with the internals of the `TCPConnection` class.

As a result of this change when writing to the connection data will be silently discarded if the connection has not yet been established.

Resolves: #1310
See also: https://github.com/ponylang/rfcs/blob/master/text/0016-tcp-must-be-connected.md