Change semantics of break with no value
break statements without a value will now generate their value from the
else branch of the loop instead of being implict break None.

The for loop has been changed. It will now break out of the loop
instead of continuing with the following iterations if Iterator.next
errors. This is for performance reasons with range check elimination.

Closes #1156.