Clarify lambda interface to map.upsert
The current example is unclear as to which value is the old value
and which is the new value when setting up the lambda. By switching
from addition to subtraction in the example, this becomes clear.

This documentation could be improved further but this is a decent
start towards resolving that possible confusion without having
to look at the source code.