Implement HMAC for crypto package
Motivation
----------
HMAC is pretty important part of crypto primitives and have to be
exposed in `crypto` package. For example to implement SCRAM SASL
authentication mechanism.

Modification
------------
New primitive has been added into `crypto` package, and has interface
similar to hash functions. It accepts name of the digest, the key and
the whole message to authenticate. Unfortunately, it is not easy to
expose chunked interface like it was done in `Digest` class, because
OpenSSL 1.0.2j (pretty widespread at the moment) does not allow to
instantiate HMAC context, i.e. does not have `HMAC_CTX_new()` function
yet.