Stop leaking memory during serialization (#1414)
Prior to this commit, serializing a pony object would leak memory.
By using "pool alloc" rather than "pony alloc", we were directly
allocating memory that would need to be manually freed. What we
wanted to do was use "pony alloc" so that the GC would track
usage of the memory and free it when needed.

Fixes #1413.