Update from_array to prevent a copy (issue #1097)
The `String.from_array` method previously was documented to only reuse
the underlying `Array[U8]` if it ended with a `0` byte, otherwise the
Array would be copied with the `0` added to the end. However, the
behavior was such that the no-copy version would make the trailing `0`
the last character of the String instead of a terminator, and changing
that behavior caused other issues.

Since pony Strings may now be non-null-terminated, the new behavior of
`.from_array` is to always use the entire underlying Array, without
regard to the presence of a trailing `0` byte.

The documentation is also updated and tests added.